fix: adjust calculation scale.

+++ src/main/java/com/test/taskTau/controller/BillsController.java
@@ -2,7 +2,12 @@ package com.test.taskTau.controller;
 
 import com.test.taskTau.service.BillsService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.math.BigDecimal;
 
 @RestController
 @RequestMapping("/bills")
@@ -12,7 +17,7 @@ public class BillsController {
     private BillsService service;
 
     @GetMapping(value = "/{bills}")
-    public int getMinimumCoins(@PathVariable int bills) throws Exception {
+    public int getMinimumCoins(@PathVariable BigDecimal bills) throws Exception {
         return service.findMiniumCoins(bills);
     }
 }
