+++ src/main/java/com/test/taskTau/enums/QuantityCoinsEnum.java
@@ -1,5 +1,7 @@
 package com.test.taskTau.enums;
 
+import java.math.BigDecimal;
+
 public enum QuantityCoinsEnum {
 
     ONE_CENT(0.01, 100),
@@ -7,11 +9,11 @@ public enum QuantityCoinsEnum {
     TEN_CENTS(0.10, 100),
     TWENTY_FIVE_CENTS(0.25, 100);
 
-    public final double value;
+    public final BigDecimal value;
     public int quantity;
 
     QuantityCoinsEnum(double value, int quantity) {
-        this.value = value;
+        this.value = BigDecimal.valueOf(value);
         this.quantity = quantity;
     }
 }
